requirements.txt
python-dateutil==2.7.3
Scrapy==1.5.1
pandas==0.23.4
numpy==1.15.4
matplotlib==3.0.2